requirements.txt
dnspython
pymongo
python-dotenv
python-telegram-bot
